Split out routes into multiple files
