+++ src/authentication/FlowDetails.tsx
@@ -15,7 +15,7 @@ import { CheckCircleIcon, TableIcon } from "@patternfly/react-icons";
 
 import type AuthenticationExecutionInfoRepresentation from "keycloak-admin/lib/defs/authenticationExecutionInfoRepresentation";
 import type AuthenticationFlowRepresentation from "keycloak-admin/lib/defs/authenticationFlowRepresentation";
-import type { FlowDetailsParams } from "./routes/Authentication";
+import type { FlowParams } from "./routes/Flow";
 import { ViewHeader } from "../components/view-header/ViewHeader";
 import { useAdminClient, useFetch } from "../context/auth/AdminClient";
 import { EmptyExecutionState } from "./EmptyExecutionState";
@@ -35,7 +35,7 @@ export const FlowDetails = () => {
   const { t } = useTranslation("authentication");
   const adminClient = useAdminClient();
   const { addAlert } = useAlerts();
-  const { id, usedBy, builtIn } = useParams<FlowDetailsParams>();
+  const { id, usedBy, builtIn } = useParams<FlowParams>();
   const [key, setKey] = useState(0);
   const refresh = () => setKey(new Date().getTime());
